savedata int->double

+++ Classes/SaveData.cpp
@@ -218,7 +218,7 @@ double SaveData::getValue(string filename, string _key, double _defaultValue)
 	string key = stringEnc(_key);
 	string v = (file_sync[file_key])[key].getString();
 	string v2 = v;//stringDecode(v);
-	int _v2 = atof(v2.c_str());
+	double _v2 = atof(v2.c_str());
 	if(v2 == "")
 		return _defaultValue;
 	else
